Renamed classes for better description
Renamed UnixPanel and WinPanel to UnixWindow and WinWindow.

Also renamed the UnixMainPanel and WinMainPanel to UnixPanel and
WinPanel.
